requirements.txt
Django==4.1.6
djangorestframework==3.14.0
pytest-django==4.5.2
psycopg2-binary==2.9.3
